+++ src/constants/storageKeys.js
@@ -1,1 +1,2 @@
 export const SNIPPET = `snippet`
+export const OPTIONS = `options`