+++ src/popup/main.js
@@ -1,13 +1,18 @@
 import Vue from 'vue'
 import Popup from './Popup'
 import '@/global.css'
+import { getOptions } from '@/lib/options'
 
-/* eslint-disable no-new */
-new Vue({
-  el: `#app`,
-  render: h => h(Popup),
-})
+(async () => {
+  const options = await getOptions()
 
-if (process.env.NODE_ENV === `development`) {
-  window.browser = browser
-}
+  /* eslint-disable no-new */
+  new Vue({
+    el: `#app`,
+    render: h => h(Popup, {
+      props: {
+        options,
+      },
+    }),
+  })
+})()